Targetable Contacts
By adding a unique id to each entry "jump to" links can be created. For
example http://localhost:9001/#c1 jumps to and highlights the second
contact.

React require unique IDs anyways, so this is an example of how that can
be utilized in the DOM also.

Why not use a CSS safe version of the contact's name? Because two
contacts could have the same name and ids must be unique.