More Traversable DOM (flex-order)
In this case we want our visual design to read contact photo, name,
address. Ideally in the DOM, the contact name would be before the image
and a heading. We can accomplish this with flexbox though!

So now the order in the DOM is
h2.contact-name
img.contact-photo
div.contact-address

which is better for search engines and screen readers but our design
still looks the same thanks to flex order.

fixing CSS syntax